+++ activerecord/test/cases/associations/eager_test.rb
@@ -61,7 +61,10 @@ class EagerAssociationTest < ActiveRecord::TestCase
   end
 
   def test_loading_conditions_with_or
-    posts = authors(:david).posts.find(:all, :include => :comments, :conditions => "comments.body like 'Normal%' OR comments.#{QUOTED_TYPE} = 'SpecialComment'")
+    posts = authors(:david).posts.find(
+      :all, :include => :comments, :references => :comments,
+      :conditions => "comments.body like 'Normal%' OR comments.#{QUOTED_TYPE} = 'SpecialComment'"
+    )
     assert_nil posts.detect { |p| p.author_id != authors(:david).id },
       "expected to find only david's posts"
   end
@@ -164,7 +167,7 @@ class EagerAssociationTest < ActiveRecord::TestCase
     car_post.categories << categories(:technology)
 
     comment = car_post.comments.create!(:body => "hmm")
-    categories = Category.find(:all, :conditions => ["posts.id=?", car_post.id],
+    categories = Category.find(:all, :conditions => { 'posts.id' => car_post.id },
                                  :include => {:posts => :comments})
     categories.each do |category|
       assert_equal [comment], category.posts[0].comments
@@ -273,17 +276,26 @@ class EagerAssociationTest < ActiveRecord::TestCase
   end
 
   def test_nested_loading_through_has_one_association_with_conditions
-    aa = AuthorAddress.find(author_addresses(:david_address).id, :include => {:author => :posts}, :conditions => "author_addresses.id > 0")
+    aa = AuthorAddress.find(
+      author_addresses(:david_address).id, :include => {:author => :posts},
+      :conditions => "author_addresses.id > 0", :references => :author_addresses
+    )
     assert_equal aa.author.posts.count, aa.author.posts.length
   end
 
   def test_nested_loading_through_has_one_association_with_conditions_on_association
-    aa = AuthorAddress.find(author_addresses(:david_address).id, :include => {:author => :posts}, :conditions => "authors.id > 0")
+    aa = AuthorAddress.find(
+      author_addresses(:david_address).id, :include => {:author => :posts},
+      :conditions => "authors.id > 0", :references => :authors
+    )
     assert_equal aa.author.posts.count, aa.author.posts.length
   end
 
   def test_nested_loading_through_has_one_association_with_conditions_on_nested_association
-    aa = AuthorAddress.find(author_addresses(:david_address).id, :include => {:author => :posts}, :conditions => "posts.id > 0")
+    aa = AuthorAddress.find(
+      author_addresses(:david_address).id, :include => {:author => :posts},
+      :conditions => "posts.id > 0", :references => :posts
+    )
     assert_equal aa.author.posts.count, aa.author.posts.length
   end
 
@@ -332,7 +344,9 @@ class EagerAssociationTest < ActiveRecord::TestCase
 
   def test_eager_association_loading_with_belongs_to_and_conditions_string_with_unquoted_table_name
     assert_nothing_raised do
-      Comment.find(:all, :include => :post, :conditions => ['posts.id = ?',4])
+      ActiveSupport::Deprecation.silence do
+        Comment.find(:all, :include => :post, :conditions => ['posts.id = ?',4])
+      end
     end
   end
 
@@ -351,7 +365,9 @@ class EagerAssociationTest < ActiveRecord::TestCase
   def test_eager_association_loading_with_belongs_to_and_conditions_string_with_quoted_table_name
     quoted_posts_id= Comment.connection.quote_table_name('posts') + '.' + Comment.connection.quote_column_name('id')
     assert_nothing_raised do
-      Comment.find(:all, :include => :post, :conditions => ["#{quoted_posts_id} = ?",4])
+      ActiveSupport::Deprecation.silence do
+        Comment.find(:all, :include => :post, :conditions => ["#{quoted_posts_id} = ?",4])
+      end
     end
   end
 
@@ -364,7 +380,9 @@ class EagerAssociationTest < ActiveRecord::TestCase
   def test_eager_association_loading_with_belongs_to_and_order_string_with_quoted_table_name
     quoted_posts_id= Comment.connection.quote_table_name('posts') + '.' + Comment.connection.quote_column_name('id')
     assert_nothing_raised do
-      Comment.find(:all, :include => :post, :order => quoted_posts_id)
+      ActiveSupport::Deprecation.silence do
+        Comment.find(:all, :include => :post, :order => quoted_posts_id)
+      end
     end
   end
 
@@ -528,22 +546,27 @@ class EagerAssociationTest < ActiveRecord::TestCase
   end
 
   def test_eager_with_has_many_and_limit_and_conditions_array_on_the_eagers
-    posts = Post.find(:all, :include => [ :author, :comments ], :limit => 2, :conditions => [ "authors.name = ?", 'David' ])
+    posts = ActiveSupport::Deprecation.silence do
+      Post.find(:all, :include => [ :author, :comments ], :limit => 2, :conditions => [ "authors.name = ?", 'David' ])
+    end
     assert_equal 2, posts.size
 
-    count = Post.count(:include => [ :author, :comments ], :limit => 2, :conditions => [ "authors.name = ?", 'David' ])
+    count = ActiveSupport::Deprecation.silence do
+      Post.count(:include => [ :author, :comments ], :limit => 2, :conditions => [ "authors.name = ?", 'David' ])
+    end
     assert_equal count, posts.size
   end
 
   def test_eager_with_has_many_and_limit_and_high_offset
-    posts = Post.find(:all, :include => [ :author, :comments ], :limit => 2, :offset => 10, :conditions => [ "authors.name = ?", 'David' ])
+    posts = Post.find(:all, :include => [ :author, :comments ], :limit => 2, :offset => 10, :conditions => { 'authors.name' => 'David' })
     assert_equal 0, posts.size
   end
 
   def test_eager_with_has_many_and_limit_and_high_offset_and_multiple_array_conditions
     assert_queries(1) do
       posts = Post.find(:all, :include => [ :author, :comments ], :limit => 2, :offset => 10,
-        :conditions => [ "authors.name = ? and comments.body = ?", 'David', 'go crazy' ])
+        :conditions => [ "authors.name = ? and comments.body = ?", 'David', 'go crazy' ],
+        :references => [:authors, :comments])
       assert_equal 0, posts.size
     end
   end
@@ -557,7 +580,7 @@ class EagerAssociationTest < ActiveRecord::TestCase
   end
 
   def test_count_eager_with_has_many_and_limit_and_high_offset
-    posts = Post.count(:all, :include => [ :author, :comments ], :limit => 2, :offset => 10, :conditions => [ "authors.name = ?", 'David' ])
+    posts = Post.count(:all, :include => [ :author, :comments ], :limit => 2, :offset => 10, :conditions => { 'authors.name' => 'David' })
     assert_equal 0, posts
   end
 
@@ -569,7 +592,7 @@ class EagerAssociationTest < ActiveRecord::TestCase
   def test_eager_count_performed_on_a_has_many_association_with_multi_table_conditional
     author = authors(:david)
     author_posts_without_comments = author.posts.select { |post| post.comments.blank? }
-    assert_equal author_posts_without_comments.size, author.posts.count(:all, :include => :comments, :conditions => 'comments.id is null')
+    assert_equal author_posts_without_comments.size, author.posts.count(:all, :include => :comments, :conditions => 'comments.id is null', :references => :comments)
   end
 
   def test_eager_count_performed_on_a_has_many_through_association_with_multi_table_conditional
@@ -608,6 +631,7 @@ class EagerAssociationTest < ActiveRecord::TestCase
     posts = authors(:david).posts.find(:all,
       :include    => :comments,
       :conditions => "comments.body like 'Normal%' OR comments.#{QUOTED_TYPE}= 'SpecialComment'",
+      :references => :comments,
       :limit      => 2
     )
     assert_equal 2, posts.size
@@ -615,6 +639,7 @@ class EagerAssociationTest < ActiveRecord::TestCase
     count = Post.count(
       :include    => [ :comments, :author ],
       :conditions => "authors.name = 'David' AND (comments.body like 'Normal%' OR comments.#{QUOTED_TYPE}= 'SpecialComment')",
+      :references => [:authors, :comments],
       :limit      => 2
     )
     assert_equal count, posts.size
@@ -624,7 +649,8 @@ class EagerAssociationTest < ActiveRecord::TestCase
     posts = nil
     Post.send(:with_scope, :find => {
       :include    => :comments,
-      :conditions => "comments.body like 'Normal%' OR comments.#{QUOTED_TYPE}= 'SpecialComment'"
+      :conditions => "comments.body like 'Normal%' OR comments.#{QUOTED_TYPE}= 'SpecialComment'",
+      :references => :comments
     }) do
       posts = authors(:david).posts.find(:all, :limit => 2)
       assert_equal 2, posts.size
@@ -632,7 +658,8 @@ class EagerAssociationTest < ActiveRecord::TestCase
 
     Post.send(:with_scope, :find => {
       :include    => [ :comments, :author ],
-      :conditions => "authors.name = 'David' AND (comments.body like 'Normal%' OR comments.#{QUOTED_TYPE}= 'SpecialComment')"
+      :conditions => "authors.name = 'David' AND (comments.body like 'Normal%' OR comments.#{QUOTED_TYPE}= 'SpecialComment')",
+      :references => [:authors, :comments]
     }) do
       count = Post.count(:limit => 2)
       assert_equal count, posts.size
@@ -644,6 +671,7 @@ class EagerAssociationTest < ActiveRecord::TestCase
       posts = authors(:david).posts.find(:all,
         :include    => :comments,
         :conditions => "comments.body like 'Normal%' OR comments.#{QUOTED_TYPE}= 'SpecialComment'",
+        :references => :comments,
         :limit      => 2
       )
       assert_equal 2, posts.size
@@ -651,6 +679,7 @@ class EagerAssociationTest < ActiveRecord::TestCase
       count = Post.count(
         :include    => [ :comments, :author ],
         :conditions => "authors.name = 'David' AND (comments.body like 'Normal%' OR comments.#{QUOTED_TYPE}= 'SpecialComment')",
+        :references => [:authors, :comments],
         :limit      => 2
       )
       assert_equal count, posts.size
@@ -658,9 +687,15 @@ class EagerAssociationTest < ActiveRecord::TestCase
   end
 
   def test_eager_with_scoped_order_using_association_limiting_without_explicit_scope
-    posts_with_explicit_order = Post.find(:all, :conditions => 'comments.id is not null', :include => :comments, :order => 'posts.id DESC', :limit => 2)
+    posts_with_explicit_order = Post.find(
+      :all, :conditions => 'comments.id is not null', :references => :comments,
+      :include => :comments, :order => 'posts.id DESC', :limit => 2
+    )
     posts_with_scoped_order = Post.send(:with_scope, :find => {:order => 'posts.id DESC'}) do
-      Post.find(:all, :conditions => 'comments.id is not null', :include => :comments, :limit => 2)
+      Post.find(
+        :all, :conditions => 'comments.id is not null',
+        :references => :comments, :include => :comments, :limit => 2
+      )
     end
     assert_equal posts_with_explicit_order, posts_with_scoped_order
   end
@@ -773,17 +808,49 @@ class EagerAssociationTest < ActiveRecord::TestCase
   end
 
   def test_limited_eager_with_order
-    assert_equal posts(:thinking, :sti_comments), Post.find(:all, :include => [:author, :comments], :conditions => "authors.name = 'David'", :order => 'UPPER(posts.title)', :limit => 2, :offset => 1)
-    assert_equal posts(:sti_post_and_comments, :sti_comments), Post.find(:all, :include => [:author, :comments], :conditions => "authors.name = 'David'", :order => 'UPPER(posts.title) DESC', :limit => 2, :offset => 1)
+    assert_equal(
+      posts(:thinking, :sti_comments),
+      Post.find(
+        :all, :include => [:author, :comments], :conditions => { 'authors.name' => 'David' },
+        :order => 'UPPER(posts.title)', :limit => 2, :offset => 1
+      )
+    )
+    assert_equal(
+      posts(:sti_post_and_comments, :sti_comments),
+      Post.find(
+        :all, :include => [:author, :comments], :conditions => { 'authors.name' => 'David' },
+        :order => 'UPPER(posts.title) DESC', :limit => 2, :offset => 1
+      )
+    )
   end
 
   def test_limited_eager_with_multiple_order_columns
-    assert_equal posts(:thinking, :sti_comments), Post.find(:all, :include => [:author, :comments], :conditions => "authors.name = 'David'", :order => ['UPPER(posts.title)', 'posts.id'], :limit => 2, :offset => 1)
-    assert_equal posts(:sti_post_and_comments, :sti_comments), Post.find(:all, :include => [:author, :comments], :conditions => "authors.name = 'David'", :order => ['UPPER(posts.title) DESC', 'posts.id'], :limit => 2, :offset => 1)
+    assert_equal(
+      posts(:thinking, :sti_comments),
+      Post.find(
+        :all, :include => [:author, :comments], :conditions => { 'authors.name' => 'David' },
+        :order => ['UPPER(posts.title)', 'posts.id'], :limit => 2, :offset => 1
+      )
+    )
+    assert_equal(
+      posts(:sti_post_and_comments, :sti_comments),
+      Post.find(
+        :all, :include => [:author, :comments], :conditions => { 'authors.name' => 'David' },
+        :order => ['UPPER(posts.title) DESC', 'posts.id'], :limit => 2, :offset => 1
+      )
+    )
   end
 
   def test_limited_eager_with_numeric_in_association
-    assert_equal people(:david, :susan), Person.find(:all, :include => [:readers, :primary_contact, :number1_fan], :conditions => "number1_fans_people.first_name like 'M%'", :order => 'people.id', :limit => 2, :offset => 0)
+    assert_equal(
+      people(:david, :susan),
+      Person.find(
+        :all, :include => [:readers, :primary_contact, :number1_fan],
+        :conditions => "number1_fans_people.first_name like 'M%'",
+        :references => :number1_fans_people,
+        :order => 'people.id', :limit => 2, :offset => 0
+      )
+    )
   end
 
   def test_preload_with_interpolation
@@ -898,11 +965,11 @@ class EagerAssociationTest < ActiveRecord::TestCase
 
   def test_count_with_include
     if current_adapter?(:SybaseAdapter)
-      assert_equal 3, authors(:david).posts_with_comments.count(:conditions => "len(comments.body) > 15")
+      assert_equal 3, authors(:david).posts_with_comments.count(:conditions => "len(comments.body) > 15", :references => :comments)
     elsif current_adapter?(:OpenBaseAdapter)
-      assert_equal 3, authors(:david).posts_with_comments.count(:conditions => "length(FETCHBLOB(comments.body)) > 15")
+      assert_equal 3, authors(:david).posts_with_comments.count(:conditions => "length(FETCHBLOB(comments.body)) > 15", :references => :comments)
     else
-      assert_equal 3, authors(:david).posts_with_comments.count(:conditions => "length(comments.body) > 15")
+      assert_equal 3, authors(:david).posts_with_comments.count(:conditions => "length(comments.body) > 15", :references => :comments)
     end
   end
 
@@ -913,7 +980,7 @@ class EagerAssociationTest < ActiveRecord::TestCase
   end
 
   def test_conditions_on_join_table_with_include_and_limit
-    assert_equal 3, Developer.find(:all, :include => 'projects', :conditions => 'developers_projects.access_level = 1', :limit => 5).size
+    assert_equal 3, Developer.find(:all, :include => 'projects', :conditions => { 'developers_projects.access_level' => 1 }, :limit => 5).size
   end
 
   def test_order_on_join_table_with_include_and_limit