+++ activerecord/test/cases/method_scoping_test.rb
@@ -104,7 +104,7 @@ class MethodScopingTest < ActiveRecord::TestCase
   def test_scoped_find_include
     # with the include, will retrieve only developers for the given project
     scoped_developers = Developer.send(:with_scope, :find => { :include => :projects }) do
-      Developer.find(:all, :conditions => 'projects.id = 2')
+      Developer.find(:all, :conditions => { 'projects.id' => 2 })
     end
     assert scoped_developers.include?(developers(:david))
     assert !scoped_developers.include?(developers(:jamis))
@@ -204,7 +204,7 @@ class MethodScopingTest < ActiveRecord::TestCase
   def test_scoped_count_include
     # with the include, will retrieve only developers for the given project
     Developer.send(:with_scope, :find => { :include => :projects }) do
-      assert_equal 1, Developer.count(:conditions => 'projects.id = 2')
+      assert_equal 1, Developer.count(:conditions => { 'projects.id' => 2 })
     end
   end
 
@@ -339,7 +339,7 @@ class NestedScopingTest < ActiveRecord::TestCase
 
   def test_nested_scoped_find_include
     Developer.send(:with_scope, :find => { :include => :projects }) do
-      Developer.send(:with_scope, :find => { :conditions => "projects.id = 2" }) do
+      Developer.send(:with_scope, :find => { :conditions => { 'projects.id' => 2 } }) do
         assert_nothing_raised { Developer.find(1) }
         assert_equal('David', Developer.find(:first).name)
       end
@@ -348,7 +348,7 @@ class NestedScopingTest < ActiveRecord::TestCase
 
   def test_nested_scoped_find_merged_include
     # :include's remain unique and don't "double up" when merging
-    Developer.send(:with_scope, :find => { :include => :projects, :conditions => "projects.id = 2" }) do
+    Developer.send(:with_scope, :find => { :include => :projects, :conditions => { 'projects.id' => 2 } }) do
       Developer.send(:with_scope, :find => { :include => :projects }) do
         assert_equal 1, Developer.scoped.includes_values.uniq.length
         assert_equal 'David', Developer.find(:first).name
@@ -356,7 +356,7 @@ class NestedScopingTest < ActiveRecord::TestCase
     end
 
     # the nested scope doesn't remove the first :include
-    Developer.send(:with_scope, :find => { :include => :projects, :conditions => "projects.id = 2" }) do
+    Developer.send(:with_scope, :find => { :include => :projects, :conditions => { 'projects.id' => 2 } }) do
       Developer.send(:with_scope, :find => { :include => [] }) do
         assert_equal 1, Developer.scoped.includes_values.uniq.length
         assert_equal('David', Developer.find(:first).name)
@@ -364,7 +364,7 @@ class NestedScopingTest < ActiveRecord::TestCase
     end
 
     # mixing array and symbol include's will merge correctly
-    Developer.send(:with_scope, :find => { :include => [:projects], :conditions => "projects.id = 2" }) do
+    Developer.send(:with_scope, :find => { :include => [:projects], :conditions => { 'projects.id' => 2 } }) do
       Developer.send(:with_scope, :find => { :include => :projects }) do
         assert_equal 1, Developer.scoped.includes_values.uniq.length
         assert_equal('David', Developer.find(:first).name)